split component tests
